requirements.txt
Django==5.0.6
gunicorn==23.0.0
djangorestframework==3.15.2
pymongo==4.7.3
django-cors-headers==4.4.0
python-decouple==3.8